requirements.txt
aiofiles==23.2.1
aiohttp==3.9.1
aioshutil==1.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
frozenlist==1.4.1
idna==3.6
multidict==6.0.4
pydantic==2.5.3
pydantic_core==2.14.6
pydash==7.0.6
PyYAML==6.0.1
sniffio==1.3.0
typing_extensions==4.9.0
websockets @ git+https://github.com/python-websockets/websockets.git@33b20e11e86f8490770185c78ed39adab8db4560
yarl==1.9.4
